Added Theme Downloader (with a catch)
Added a "Fetch Themes..." option to the theme selector, which if selected will fetch all GUIs mods from GameBanana and add them to the list. clicking on any of the new entries will automatically download and install it

The catch? I haven't tested it in a real scenario yet because GetByNameAsync does not return

Also I removed the luna theme, I really liked that one but Microsoft might have something to say about it so it's gone now